# src/Routes/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "Documentation HML THANOS"
    },
...
    "paths": {
...
        "/descargar-data/products-approveds": {
            "post": {
                "tags": ["Download Data"],
                "summary": "Download file products approveds",
                "responses": {
                    "200": {
                        "description": "El archivo se descargó correctamente",
                        "content" : {
                            "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet" : {
                                "schema" : {
                                    "type" : "object",
                                    "format" : "binary",
                                    "xml" : {
                                        "name" : "datos.xlsx"
                                    }
                                }
                            }
                        }
                    },
                    "500" : {
                        "description" : "Lo sentimos hubo un error al momento de descargar los productos homologados",
                        "content" : {
                            "application/json" : {
                                "schema" : {
                                    "type" : "object",
                                    "properties" : {
                                        "message": {
                                            "type": "string",
                                            "example" :  "Lo sentimos hubo un error al momento de descargar los productos homologados"
                                        },
                                        "devmsg": {
                                            "type": "array",
                                            "example" : ["Error server"]
                                        },
                                        "respuesta": {
                                            "type": "boolean",
                                            "example" : false
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/descargar-data/distributors": {
            "post": {
                "tags": ["Download Data"],
                "summary": "Get url download file distribuitors",
                "responses": {
                    "200": {
                        "description": "Se descargó exitosamente",
                        "content" : {
                            "application/json" : {
                                "schema" : {
                                    "type" : "object",
                                    "properties" : {
                                        "message": {
                                            "type": "string",
                                            "example" : "Se descargó exitosamente."
                                        },
                                        "respuesta": {
                                            "type": "boolean",
                                            "example" : true
                                        },
                                        "data": {
                                            "type": "string",
                                            "example" : "WwwUrlDistribuidorsCom"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500" : {
                        "description" : "Lo sentimos hubo un error al momento de descargar.",
                        "content" : {
                            "application/json" : {
                                "schema" : {
                                    "type" : "object",
                                    "properties" : {
                                        "message": {
                                            "type": "string",
                                            "example" :  "Lo sentimos hubo un error al momento de descargar."
                                        },
                                        "devmsg": {
                                            "type": "array",
                                            "example" : ["Error server"]
                                        },
                                        "respuesta": {
                                            "type": "boolean",
                                            "example" : false
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/descargar-data/products": {
            "post": {
                "tags": ["Download Data"],
                "summary": "Get url download file products",
                "responses": {
                    "200": {
                        "description": "Se descargó exitosamente",
                        "content" : {
                            "application/json" : {
                                "schema" : {
                                    "type" : "object",
                                    "properties" : {
                                        "message": {
                                            "type": "string",
                                            "example" : "Se descargó exitosamente."
                                        },
                                        "respuesta": {
                                            "type": "boolean",
                                            "example" : true
                                        },
                                        "data": {
                                            "type": "string",
                                            "example" : "WwwUrlProductsCom"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500" : {
                        "description" : "Lo sentimos hubo un error al momento de descargar.",
                        "content" : {
                            "application/json" : {
                                "schema" : {
                                    "type" : "object",
                                    "properties" : {
                                        "message": {
                                            "type": "string",
                                            "example" :  "Lo sentimos hubo un error al momento de descargar."
                                        },
                                        "devmsg": {
                                            "type": "array",
                                            "example" : ["Error server"]
                                        },
                                        "respuesta": {
                                            "type": "boolean",
                                            "example" : false
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/descargar-data/products-so": {
            "post": {
                "tags": ["Download Data"],
                "summary": "Get url download file products approved",
                "responses": {
                    "200": {
                        "description": "Se descargó exitosamente",
                        "content" : {
                            "application/json" : {
                                "schema" : {
                                    "type" : "object",
                                    "properties" : {
                                        "message": {
                                            "type": "string",
                                            "example" : "Se descargó exitosamente."
                                        },
                                        "respuesta": {
                                            "type": "boolean",
                                            "example" : true
                                        },
                                        "data": {
                                            "type": "string",
                                            "example" : "WwwUrlProductsApprovedCom"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500" : {
                        "description" : "Lo sentimos hubo un error al momento de descargar.",
                        "content" : {
                            "application/json" : {
                                "schema" : {
                                    "type" : "object",
                                    "properties" : {
                                        "message": {
                                            "type": "string",
                                            "example" :  "Lo sentimos hubo un error al momento de descargar."
                                        },
                                        "devmsg": {
                                            "type": "array",
                                            "example" : ["Error server"]
                                        },
                                        "respuesta": {
                                            "type": "boolean",
                                            "example" : false
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/descargar-data/download-file/{type}/{namefile}": {
            "get": {
                "tags": ["Download Data"],
                "summary": "Download file",
                "responses": {
                    "200": {
                        "description": "El archivo se descargó correctamente"
                    },
                    "500" : {
                        "description" : "Ocurrió un error al descargar el archivo."
                    }
                }
            }
        },
...
    }
}